+++ CHANGELOG.md
@@ -37,6 +37,9 @@
 + [x] Input: add style prop `min-width=0`, issue #1626
 + [x] Component: fix method `_runtime()` for updating attribute `data-role`.
 + [x] Input: add attribute `data-label`. If this attribute defined, `label` element for input will be created automatically.
++ [x] Calendar picker: add attribute `data-label`. If this attribute defined, `label` element for input will be created automatically.
++ [x] Time picker: add attribute `data-label`. If this attribute defined, `label` element for input will be created automatically.
++ [x] Date picker: add attribute `data-label`. If this attribute defined, `label` element for input will be created automatically.
 
 ### 4.3.10
 + [x] Input: add API method `setAutocompleteList(array|string)`, issue #1576